backend/requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
distro==1.9.0
dnspython==2.8.0
dotenv==0.9.9
ecdsa==0.19.2
email-validator==2.3.0
fastapi==0.136.0
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
jiter==0.14.0
jsonpatch==1.33
jsonpointer==3.1.1
langchain-core==1.3.0
langchain-openai==1.1.16
langgraph==1.1.9
langgraph-checkpoint==4.0.2
langgraph-prebuilt==1.0.10
langgraph-sdk==0.3.13
langsmith==0.7.33
Mako==1.3.11
MarkupSafe==3.0.3
openai==2.32.0
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.1
passlib==1.7.4
pluggy==1.6.0
psycopg2-binary==2.9.12
pyasn1==0.6.3
pydantic==2.13.3
pydantic-settings==2.14.0
pydantic_core==2.46.3
Pygments==2.20.0
pytest==9.0.3
pytest-asyncio==1.3.0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.26
pypdf==5.4.0
python-docx==1.1.2
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.49
starlette==1.0.0
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.45.0
xxhash==3.6.0
zstandard==0.25.0
